+++ src/ops/disting.c
@@ -58,11 +58,17 @@ static void op_EX_M_CH_set(const void *data, scene_state_t *ss,
                            exec_state_t *es, command_state_t *cs);
 static void op_EX_M_N_get(const void *data, scene_state_t *ss, exec_state_t *es,
                           command_state_t *cs);
+static void op_EX_M_N_POUND_get(const void *data, scene_state_t *ss,
+                                exec_state_t *es, command_state_t *cs);
 static void op_EX_M_NO_get(const void *data, scene_state_t *ss,
                            exec_state_t *es, command_state_t *cs);
-static void op_EX_M_PB_get(const void *data, scene_state_t *ss,
-                           exec_state_t *es, command_state_t *cs);
+static void op_EX_M_NO_POUND_get(const void *data, scene_state_t *ss,
+                                 exec_state_t *es, command_state_t *cs);
 static void op_EX_M_CC_get(const void *data, scene_state_t *ss,
+                           exec_state_t *es, command_state_t *cs);
+static void op_EX_M_CC_POUND_get(const void *data, scene_state_t *ss,
+                                 exec_state_t *es, command_state_t *cs);
+static void op_EX_M_PB_get(const void *data, scene_state_t *ss,
                            exec_state_t *es, command_state_t *cs);
 static void op_EX_M_PRG_get(const void *data, scene_state_t *ss,
                             exec_state_t *es, command_state_t *cs);
@@ -128,6 +134,57 @@ static void op_EX_LP_REVQ_get(const void *data, scene_state_t *ss,
                               exec_state_t *es, command_state_t *cs);
 static void op_EX_LP_DOWNQ_get(const void *data, scene_state_t *ss,
                                exec_state_t *es, command_state_t *cs);
+static void op_EX_A1_get(const void *data, scene_state_t *ss, exec_state_t *es,
+                         command_state_t *cs);
+static void op_EX_A1_set(const void *data, scene_state_t *ss, exec_state_t *es,
+                         command_state_t *cs);
+static void op_EX_A2_get(const void *data, scene_state_t *ss, exec_state_t *es,
+                         command_state_t *cs);
+static void op_EX_A2_set(const void *data, scene_state_t *ss, exec_state_t *es,
+                         command_state_t *cs);
+static void op_EX_A12_get(const void *data, scene_state_t *ss, exec_state_t *es,
+                          command_state_t *cs);
+static void op_EX_P1_get(const void *data, scene_state_t *ss, exec_state_t *es,
+                         command_state_t *cs);
+static void op_EX_P1_set(const void *data, scene_state_t *ss, exec_state_t *es,
+                         command_state_t *cs);
+static void op_EX_P2_get(const void *data, scene_state_t *ss, exec_state_t *es,
+                         command_state_t *cs);
+static void op_EX_P2_set(const void *data, scene_state_t *ss, exec_state_t *es,
+                         command_state_t *cs);
+static void op_EX_PV1_get(const void *data, scene_state_t *ss, exec_state_t *es,
+                          command_state_t *cs);
+static void op_EX_PV2_get(const void *data, scene_state_t *ss, exec_state_t *es,
+                          command_state_t *cs);
+static void op_EX_MIN1_get(const void *data, scene_state_t *ss,
+                           exec_state_t *es, command_state_t *cs);
+static void op_EX_MIN2_get(const void *data, scene_state_t *ss,
+                           exec_state_t *es, command_state_t *cs);
+static void op_EX_MAX1_get(const void *data, scene_state_t *ss,
+                           exec_state_t *es, command_state_t *cs);
+static void op_EX_MAX2_get(const void *data, scene_state_t *ss,
+                           exec_state_t *es, command_state_t *cs);
+static void op_EX_PRE1_get(const void *data, scene_state_t *ss,
+                           exec_state_t *es, command_state_t *cs);
+static void op_EX_PRE2_get(const void *data, scene_state_t *ss,
+                           exec_state_t *es, command_state_t *cs);
+static void op_EX_SAVE1_get(const void *data, scene_state_t *ss,
+                            exec_state_t *es, command_state_t *cs);
+static void op_EX_SAVE2_get(const void *data, scene_state_t *ss,
+                            exec_state_t *es, command_state_t *cs);
+static void op_EX_Z1_get(const void *data, scene_state_t *ss, exec_state_t *es,
+                         command_state_t *cs);
+static void op_EX_Z1_set(const void *data, scene_state_t *ss, exec_state_t *es,
+                         command_state_t *cs);
+static void op_EX_Z2_get(const void *data, scene_state_t *ss, exec_state_t *es,
+                         command_state_t *cs);
+static void op_EX_Z2_set(const void *data, scene_state_t *ss, exec_state_t *es,
+                         command_state_t *cs);
+static void op_EX_ZO1_get(const void *data, scene_state_t *ss, exec_state_t *es,
+                          command_state_t *cs);
+static void op_EX_ZO2_get(const void *data, scene_state_t *ss, exec_state_t *es,
+                          command_state_t *cs);
+
 // clang-format off
                    
 const tele_mod_t mod_EX1       = MAKE_MOD(EX1, mod_EX1_func, 0);
@@ -150,16 +207,19 @@ const tele_op_t op_EX_PLAY     = MAKE_GET_OP(EX.PLAY,       op_EX_PLAY_get,     
 const tele_op_t op_EX_AL_P     = MAKE_GET_OP(EX.AL.P,       op_EX_AL_P_get,     1, false);
 const tele_op_t op_EX_AL_CLK   = MAKE_GET_OP(EX.AL.CLK,     op_EX_AL_CLK_get,   0, false);
 
-const tele_op_t op_EX_M_CH     = MAKE_GET_SET_OP(EX.M.CH,   op_EX_M_CH_get,     op_EX_M_CH_set, 0, true);
-const tele_op_t op_EX_M_N      = MAKE_GET_OP(EX.M.N,        op_EX_M_N_get,      2, false);
-const tele_op_t op_EX_M_NO     = MAKE_GET_OP(EX.M.NO,       op_EX_M_NO_get,     1, false);
-const tele_op_t op_EX_M_PB     = MAKE_GET_OP(EX.M.PB,       op_EX_M_PB_get,     1, false);
-const tele_op_t op_EX_M_CC     = MAKE_GET_OP(EX.M.CC,       op_EX_M_CC_get,     2, false);
-const tele_op_t op_EX_M_PRG    = MAKE_GET_OP(EX.M.PRG,      op_EX_M_PRG_get,    1, false);
-const tele_op_t op_EX_M_CLK    = MAKE_GET_OP(EX.M.CLK,      op_EX_M_CLK_get,    0, false);
-const tele_op_t op_EX_M_START  = MAKE_GET_OP(EX.M.START,    op_EX_M_START_get,  0, false);
-const tele_op_t op_EX_M_STOP   = MAKE_GET_OP(EX.M.STOP,     op_EX_M_STOP_get,   0, false);
-const tele_op_t op_EX_M_CONT   = MAKE_GET_OP(EX.M.CONT,     op_EX_M_CONT_get,   0, false);
+const tele_op_t op_EX_M_CH       = MAKE_GET_SET_OP(EX.M.CH, op_EX_M_CH_get, op_EX_M_CH_set, 0, true);
+const tele_op_t op_EX_M_N        = MAKE_GET_OP(EX.M.N,      op_EX_M_N_get,        2, false);
+const tele_op_t op_EX_M_N_POUND  = MAKE_GET_OP(EX.M.N#,     op_EX_M_N_POUND_get,  3, false);
+const tele_op_t op_EX_M_NO       = MAKE_GET_OP(EX.M.NO,     op_EX_M_NO_get,       1, false);
+const tele_op_t op_EX_M_NO_POUND = MAKE_GET_OP(EX.M.NO#,    op_EX_M_NO_POUND_get, 2, false);
+const tele_op_t op_EX_M_CC       = MAKE_GET_OP(EX.M.CC,     op_EX_M_CC_get,       2, false);
+const tele_op_t op_EX_M_CC_POUND = MAKE_GET_OP(EX.M.CC#,    op_EX_M_CC_POUND_get, 3, false);
+const tele_op_t op_EX_M_PB       = MAKE_GET_OP(EX.M.PB,     op_EX_M_PB_get,       1, false);
+const tele_op_t op_EX_M_PRG      = MAKE_GET_OP(EX.M.PRG,    op_EX_M_PRG_get,      1, false);
+const tele_op_t op_EX_M_CLK      = MAKE_GET_OP(EX.M.CLK,    op_EX_M_CLK_get,      0, false);
+const tele_op_t op_EX_M_START    = MAKE_GET_OP(EX.M.START,  op_EX_M_START_get,    0, false);
+const tele_op_t op_EX_M_STOP     = MAKE_GET_OP(EX.M.STOP,   op_EX_M_STOP_get,     0, false);
+const tele_op_t op_EX_M_CONT     = MAKE_GET_OP(EX.M.CONT,   op_EX_M_CONT_get,     0, false);
 
 const tele_op_t op_EX_SB_CH    = MAKE_GET_SET_OP(EX.SB.CH,  op_EX_SB_CH_get,    op_EX_SB_CH_set, 0, true);
 const tele_op_t op_EX_SB_N     = MAKE_GET_OP(EX.SB.N,       op_EX_SB_N_get,     2, false);
@@ -187,6 +247,26 @@ const tele_op_t op_EX_LP_CLR   = MAKE_GET_OP(EX.LP.CLR,   op_EX_LP_CLR_get,   1,
 const tele_op_t op_EX_LP       = MAKE_GET_OP(EX.LP,       op_EX_LP_get,       1, true);
 const tele_op_t op_EX_LP_REVQ  = MAKE_GET_OP(EX.LP.REV?,  op_EX_LP_REVQ_get,  1, true);
 const tele_op_t op_EX_LP_DOWNQ = MAKE_GET_OP(EX.LP.DOWN?, op_EX_LP_DOWNQ_get, 1, true);
+
+const tele_op_t op_EX_A1       = MAKE_GET_SET_OP(EX.A1, op_EX_A1_get,    op_EX_A1_set, 0, true);
+const tele_op_t op_EX_A2       = MAKE_GET_SET_OP(EX.A2, op_EX_A2_get,    op_EX_A2_set, 0, true);
+const tele_op_t op_EX_A12      = MAKE_GET_OP(EX.A12,    op_EX_A12_get,   2, false);
+const tele_op_t op_EX_P1       = MAKE_GET_SET_OP(EX.P1, op_EX_P1_get,    op_EX_P1_set, 1, true);
+const tele_op_t op_EX_P2       = MAKE_GET_SET_OP(EX.P2, op_EX_P2_get,    op_EX_P2_set, 1, true);
+const tele_op_t op_EX_PV1      = MAKE_GET_OP(EX.PV1,    op_EX_PV1_get,   2, false);
+const tele_op_t op_EX_PV2      = MAKE_GET_OP(EX.PV2,    op_EX_PV2_get,   2, false);
+const tele_op_t op_EX_MIN1     = MAKE_GET_OP(EX.MIN1,   op_EX_MIN1_get,  1, true);
+const tele_op_t op_EX_MIN2     = MAKE_GET_OP(EX.MIN2,   op_EX_MIN2_get,  1, true);
+const tele_op_t op_EX_MAX1     = MAKE_GET_OP(EX.MAX1,   op_EX_MAX1_get,  1, true);
+const tele_op_t op_EX_MAX2     = MAKE_GET_OP(EX.MAX2,   op_EX_MAX2_get,  1, true);
+const tele_op_t op_EX_PRE1     = MAKE_GET_OP(EX.PRE1,   op_EX_PRE1_get,  1, false);
+const tele_op_t op_EX_PRE2     = MAKE_GET_OP(EX.PRE2,   op_EX_PRE2_get,  1, false);
+const tele_op_t op_EX_SAVE1    = MAKE_GET_OP(EX.SAVE1,  op_EX_SAVE1_get, 1, false);
+const tele_op_t op_EX_SAVE2    = MAKE_GET_OP(EX.SAVE2,  op_EX_SAVE2_get, 1, false);
+const tele_op_t op_EX_Z1       = MAKE_GET_SET_OP(EX.Z1, op_EX_Z1_get,    op_EX_Z1_set, 0, true);
+const tele_op_t op_EX_Z2       = MAKE_GET_SET_OP(EX.Z2, op_EX_Z2_get,    op_EX_Z2_set, 0, true);
+const tele_op_t op_EX_ZO1      = MAKE_GET_OP(EX.ZO1,    op_EX_ZO1_get,   0, false);
+const tele_op_t op_EX_ZO2      = MAKE_GET_OP(EX.ZO2,    op_EX_ZO2_get,   0, false);
 
 const tele_op_t op_EX_PRE = MAKE_ALIAS_OP(EX.PRE, op_EX_PRESET_get, op_EX_PRESET_set, 0, true);
 const tele_op_t op_EX_A   = MAKE_ALIAS_OP(EX.A,   op_EX_ALG_get,    op_EX_ALG_set,    0, true);
@@ -420,6 +500,18 @@ static void op_EX_M_N_get(const void *NOTUSED(data), scene_state_t *ss,
     send4(0x4F, 0x90 + midi_channel, note, velocity);
 }
 
+static void op_EX_M_N_POUND_get(const void *NOTUSED(data), scene_state_t *ss,
+                                exec_state_t *NOTUSED(es),
+                                command_state_t *cs) {
+    s16 ch = cs_pop(cs) - 1;
+    u16 note = cs_pop(cs);
+    u16 velocity = cs_pop(cs);
+    if (ch < 0 || ch > 15) return;
+    if (note > 127) return;
+    if (velocity > 127) velocity = 127;
+    send4(0x4F, 0x90 + ch, note, velocity);
+}
+
 static void op_EX_M_NO_get(const void *NOTUSED(data), scene_state_t *ss,
                            exec_state_t *NOTUSED(es), command_state_t *cs) {
     u16 note = cs_pop(cs);
@@ -427,10 +519,14 @@ static void op_EX_M_NO_get(const void *NOTUSED(data), scene_state_t *ss,
     send4(0x4F, 0x80 + midi_channel, note, 0);
 }
 
-static void op_EX_M_PB_get(const void *NOTUSED(data), scene_state_t *ss,
-                           exec_state_t *NOTUSED(es), command_state_t *cs) {
-    u16 bend = cs_pop(cs);
-    send4(0x4F, 0xE0 + midi_channel, bend, bend >> 8);
+static void op_EX_M_NO_POUND_get(const void *NOTUSED(data), scene_state_t *ss,
+                                 exec_state_t *NOTUSED(es),
+                                 command_state_t *cs) {
+    s16 ch = cs_pop(cs) - 1;
+    u16 note = cs_pop(cs);
+    if (ch < 0 || ch > 15) return;
+    if (note > 127) return;
+    send4(0x4F, 0x80 + ch, note, 0);
 }
 
 static void op_EX_M_CC_get(const void *NOTUSED(data), scene_state_t *ss,
@@ -440,6 +536,24 @@ static void op_EX_M_CC_get(const void *NOTUSED(data), scene_state_t *ss,
     if (controller > 127) return;
     if (value > 127) value = 127;
     send4(0x4F, 0xB0 + midi_channel, controller, value);
+}
+
+static void op_EX_M_CC_POUND_get(const void *NOTUSED(data), scene_state_t *ss,
+                                 exec_state_t *NOTUSED(es),
+                                 command_state_t *cs) {
+    s16 ch = cs_pop(cs) - 1;
+    u16 controller = cs_pop(cs);
+    u16 value = cs_pop(cs);
+    if (ch < 0 || ch > 15) return;
+    if (controller > 127) return;
+    if (value > 127) value = 127;
+    send4(0x4F, 0xB0 + ch, controller, value);
+}
+
+static void op_EX_M_PB_get(const void *NOTUSED(data), scene_state_t *ss,
+                           exec_state_t *NOTUSED(es), command_state_t *cs) {
+    u16 bend = cs_pop(cs);
+    send4(0x4F, 0xE0 + midi_channel, bend, bend >> 8);
 }
 
 static void op_EX_M_PRG_get(const void *NOTUSED(data), scene_state_t *ss,
@@ -713,4 +827,206 @@ static void op_EX_LP_DOWNQ_get(const void *NOTUSED(data), scene_state_t *ss,
     }
 
     cs_push(cs, get_looper_state(loop) & 0b100000 ? 1 : 0);
+}
+
+static void op_EX_A1_get(const void *NOTUSED(data), scene_state_t *NOTUSED(ss),
+                         exec_state_t *NOTUSED(es), command_state_t *cs) {
+    send2(0x5F, 0);
+    data[0] = 0;
+    tele_ii_rx(DISTING_EX_1 + unit, data, 1);
+    cs_push(cs, data[0] + 1);
+}
+
+static void op_EX_A1_set(const void *NOTUSED(data), scene_state_t *NOTUSED(ss),
+                         exec_state_t *NOTUSED(es), command_state_t *cs) {
+    u16 algo = cs_pop(cs);
+    if (algo < 1) return;
+    send3(0x60, 0, algo - 1);
+}
+
+static void op_EX_A2_get(const void *NOTUSED(data), scene_state_t *NOTUSED(ss),
+                         exec_state_t *NOTUSED(es), command_state_t *cs) {
+    send2(0x5F, 1);
+    data[0] = 0;
+    tele_ii_rx(DISTING_EX_1 + unit, data, 1);
+    cs_push(cs, data[0] + 1);
+}
+
+static void op_EX_A2_set(const void *NOTUSED(data), scene_state_t *NOTUSED(ss),
+                         exec_state_t *NOTUSED(es), command_state_t *cs) {
+    u16 algo = cs_pop(cs);
+    if (algo < 1) return;
+    send3(0x60, 1, algo - 1);
+}
+
+static void op_EX_A12_get(const void *NOTUSED(data), scene_state_t *NOTUSED(ss),
+                          exec_state_t *NOTUSED(es), command_state_t *cs) {
+    u16 algo1 = cs_pop(cs);
+    u16 algo2 = cs_pop(cs);
+    if (algo1 < 1 || algo2 < 1) return;
+    send3(0x62, algo1 - 1, algo2 - 1);
+}
+
+static void op_EX_P1_get(const void *NOTUSED(data), scene_state_t *NOTUSED(ss),
+                         exec_state_t *NOTUSED(es), command_state_t *cs) {
+    u16 param = cs_pop(cs);
+    send2(0x5A, param);
+
+    data[0] = 0;
+    tele_ii_rx(DISTING_EX_1 + unit, data, 1);
+    cs_push(cs, (s8)data[0]);
+}
+
+static void op_EX_P1_set(const void *NOTUSED(data), scene_state_t *NOTUSED(ss),
+                         exec_state_t *NOTUSED(es), command_state_t *cs) {
+    u16 param = cs_pop(cs);
+    u16 value = cs_pop(cs);
+    send3(0x5D, param, value);
+}
+
+static void op_EX_P2_get(const void *NOTUSED(data), scene_state_t *NOTUSED(ss),
+                         exec_state_t *NOTUSED(es), command_state_t *cs) {
+    u16 param = cs_pop(cs);
+    send2(0x5A, param | 0b10000);
+
+    data[0] = 0;
+    tele_ii_rx(DISTING_EX_1 + unit, data, 1);
+    cs_push(cs, (s8)data[0]);
+}
+
+static void op_EX_P2_set(const void *NOTUSED(data), scene_state_t *NOTUSED(ss),
+                         exec_state_t *NOTUSED(es), command_state_t *cs) {
+    u16 param = cs_pop(cs);
+    u16 value = cs_pop(cs);
+    send3(0x5D, param | 0b10000, value);
+}
+
+static void op_EX_PV1_get(const void *NOTUSED(data), scene_state_t *NOTUSED(ss),
+                          exec_state_t *NOTUSED(es), command_state_t *cs) {
+    u16 param = cs_pop(cs);
+    u16 value = cs_pop(cs);
+    send4(0x5E, param, value >> 8, value);
+}
+
+static void op_EX_PV2_get(const void *NOTUSED(data), scene_state_t *NOTUSED(ss),
+                          exec_state_t *NOTUSED(es), command_state_t *cs) {
+    u16 param = cs_pop(cs);
+    u16 value = cs_pop(cs);
+    send4(0x5E, param | 0b10000, value >> 8, value);
+}
+
+static void op_EX_MIN1_get(const void *NOTUSED(data),
+                           scene_state_t *NOTUSED(ss),
+                           exec_state_t *NOTUSED(es), command_state_t *cs) {
+    u16 param = cs_pop(cs);
+    send2(0x5B, param);
+
+    data[0] = 0;
+    tele_ii_rx(DISTING_EX_1 + unit, data, 1);
+    cs_push(cs, (s8)data[0]);
+}
+
+static void op_EX_MIN2_get(const void *NOTUSED(data),
+                           scene_state_t *NOTUSED(ss),
+                           exec_state_t *NOTUSED(es), command_state_t *cs) {
+    u16 param = cs_pop(cs);
+    send2(0x5B, param | 0b10000);
+
+    data[0] = 0;
+    tele_ii_rx(DISTING_EX_1 + unit, data, 1);
+    cs_push(cs, (s8)data[0]);
+}
+
+static void op_EX_MAX1_get(const void *NOTUSED(data),
+                           scene_state_t *NOTUSED(ss),
+                           exec_state_t *NOTUSED(es), command_state_t *cs) {
+    u16 param = cs_pop(cs);
+    send2(0x5C, param);
+
+    data[0] = 0;
+    tele_ii_rx(DISTING_EX_1 + unit, data, 1);
+    cs_push(cs, (s8)data[0]);
+}
+
+static void op_EX_MAX2_get(const void *NOTUSED(data),
+                           scene_state_t *NOTUSED(ss),
+                           exec_state_t *NOTUSED(es), command_state_t *cs) {
+    u16 param = cs_pop(cs);
+    send2(0x5C, param | 0b10000);
+
+    data[0] = 0;
+    tele_ii_rx(DISTING_EX_1 + unit, data, 1);
+    cs_push(cs, (s8)data[0]);
+}
+
+static void op_EX_PRE1_get(const void *NOTUSED(data),
+                           scene_state_t *NOTUSED(ss),
+                           exec_state_t *NOTUSED(es), command_state_t *cs) {
+    u16 preset = cs_pop(cs);
+    send3(0x63, 0, preset);
+}
+
+static void op_EX_PRE2_get(const void *NOTUSED(data),
+                           scene_state_t *NOTUSED(ss),
+                           exec_state_t *NOTUSED(es), command_state_t *cs) {
+    u16 preset = cs_pop(cs);
+    send3(0x63, 1, preset);
+}
+
+static void op_EX_SAVE1_get(const void *NOTUSED(data),
+                            scene_state_t *NOTUSED(ss),
+                            exec_state_t *NOTUSED(es), command_state_t *cs) {
+    u16 preset = cs_pop(cs);
+    send3(0x64, 0, preset);
+}
+
+static void op_EX_SAVE2_get(const void *NOTUSED(data),
+                            scene_state_t *NOTUSED(ss),
+                            exec_state_t *NOTUSED(es), command_state_t *cs) {
+    u16 preset = cs_pop(cs);
+    send3(0x64, 1, preset);
+}
+
+static void op_EX_Z1_get(const void *NOTUSED(data), scene_state_t *NOTUSED(ss),
+                         exec_state_t *NOTUSED(es), command_state_t *cs) {
+    send1(0x66);
+
+    data[0] = data[1] = data[2] = data[3] = 0;
+    tele_ii_rx(DISTING_EX_1 + unit, data, 4);
+    u16 value = (data[0] << 8) + data[1];
+    cs_push(cs, value >> 8);
+}
+
+static void op_EX_Z1_set(const void *NOTUSED(data), scene_state_t *NOTUSED(ss),
+                         exec_state_t *NOTUSED(es), command_state_t *cs) {
+    u16 param = cs_pop(cs);
+    if (param > 127) return;
+    send3(0x65, 0, param);
+}
+
+static void op_EX_Z2_get(const void *NOTUSED(data), scene_state_t *NOTUSED(ss),
+                         exec_state_t *NOTUSED(es), command_state_t *cs) {
+    send1(0x66);
+
+    data[0] = data[1] = data[2] = data[3] = 0;
+    tele_ii_rx(DISTING_EX_1 + unit, data, 4);
+    u16 value = (data[2] << 8) + data[3];
+    cs_push(cs, (s16)value >> 8);
+}
+
+static void op_EX_Z2_set(const void *NOTUSED(data), scene_state_t *NOTUSED(ss),
+                         exec_state_t *NOTUSED(es), command_state_t *cs) {
+    u16 param = cs_pop(cs);
+    if (param > 127) return;
+    send3(0x65, 1, param);
+}
+
+static void op_EX_ZO1_get(const void *NOTUSED(data), scene_state_t *NOTUSED(ss),
+                          exec_state_t *NOTUSED(es), command_state_t *cs) {
+    send3(0x65, 0, 128);
+}
+
+static void op_EX_ZO2_get(const void *NOTUSED(data), scene_state_t *NOTUSED(ss),
+                          exec_state_t *NOTUSED(es), command_state_t *cs) {
+    send3(0x65, 1, 128);
 }